Silly Ubuntu

+++ src/tool/abi/main.cpp
@@ -301,7 +301,7 @@ namespace ABI {
 
         if (config.verbose)
         {
-            w.write("time: %ms\n", duration_cast<milliseconds>((high_resolution_clock::now() - start)).count());
+            w.write("time: %ms\n", static_cast<std::int64_t>(duration_cast<milliseconds>((high_resolution_clock::now() - start)).count()));
         }
     }
     catch (std::exception const& e)
